requirements.txt
pytorch==1.9.0
opencv-python==4.5.2.54
h5py==2.10.0
scikit-learn==0.24.2
pillow==7.1.2
scipy==1.7.0
numpy==1.9.2